organize linters in separate folder
