+++ spec/linters/coala/coala_json.json
