+++ spec/spec_helper.rb
@@ -1,0 +1,3 @@
+require_relative "../src/linters/coala/base_linter"
+require 'rspec'
+require 'json'
