+++ src/linters/coala/base_linter.rb
