+++ src/linters/coala/java_linter.rb
